Fixed Issue 116 Changed volumes from double to float, because in the end we need float, because that's what the Android API wants. See http://developer.android.com/reference/android/media/MediaPlayer.html#setVolume(float, float)
Also, I was unable to think of a test for this, because while we can _set_ the volume of individual MediaPlayers, there is currently no way of getting it, besides remembering it.

Update issue 116
Owner: aried3r@gmail.com
I fixed this issue, note however that the volume is a global one amongst all media players. When setting the volume to 100%, _all_ sounds in _all_ sprites will have 100%.

+++ catroid/src/at/tugraz/ist/catroid/content/bricks/ChangeVolumeByBrick.java
@@ -55,12 +55,12 @@ public class ChangeVolumeByBrick implements Brick, OnClickListener {
 	}
 
 	public void execute() {
-		double currentVolume = SoundManager.getInstance().getVolume();
+		float currentVolume = SoundManager.getInstance().getVolume();
 		currentVolume += volume;
-		if (currentVolume < 0.0) {
-			currentVolume = 0.0;
-		} else if (currentVolume > 100.0) {
-			currentVolume = 100.0;
+		if (currentVolume < 0.0f) {
+			currentVolume = 0.0f;
+		} else if (currentVolume > 100.0f) {
+			currentVolume = 100.0f;
 		}
 		SoundManager.getInstance().setVolume(currentVolume);
 	}
